AP-17403: Make matching of some Container Input schemas more precise
These Schemas should not match an empty JSON document. This is because
the new Container Input (Variable) can be set to accept any schema and
provides an empty JSON in that case. See ticket for more background.

Additionall, adjust visibility of ContainerVariableMapper2 to be
available for node implementations.

AP-17403 (Call workflow table-based does not recognise Container input variable)