AP-22461: use JSONCellContent more consistently (avoid parsing)
This change is most significant in terms of performance. It avoids unnecessary
creation of jakarta.json.JsonValue, which is done lazy in the cell
implementation, ... just to take the JsonValue and write it to a String again
for persistance.

AP-22461 (Columnar Backend: Improve Json cell handling by avoiding unnecessary (un)marshalling)